mostly filled in result dao
Former-commit-id: c39b49ab31a9daa5cc52cf3f60d3c4b9a231fdb8 [formerly 51f7977804661b94745b4fd277038b06e2e78ae3]
Former-commit-id: 9a2f632bb4559947dbb3d69c5d1fd03b0c7334e9

+++ src/mitll/langtest/client/LangTestDatabaseAsync.java
@@ -138,7 +138,7 @@ public interface LangTestDatabaseAsync {
 
   void addRoundTrip(long resultid, int roundTrip, AsyncCallback<Void> async);
 
-  void getResultASRInfo(long resultID, int width, int height, AsyncCallback<PretestScore> async);
+  void getResultASRInfo(int resultID, int width, int height, AsyncCallback<PretestScore> async);
 
 
   void getASRScoreForAudioPhonemes(int reqid, long resultID, String testAudioFile, String sentence,
